requirements.txt
fire
numpy
pandas
diskcache
pandarallel
geoparse
# torch
torchmetrics
pytorch_lightning